+++ app/Http/Controllers/Admin/RoleController.php
@@ -20,6 +20,8 @@ class RoleController extends Controller
 {
     public function index(Request $request): View | JsonResponse
     {
+        validate_permission('roles.read');
+
         if ($request->ajax()) {
             $roles = Role::offset($request->start)->limit($request->length);
             $totalRecords = Role::count();
@@ -30,8 +32,12 @@ class RoleController extends Controller
                 ->addColumn('actions', function ($role) {
                     return Blade::render('
                         <div class="btn-group">
-                            <a href="{{ route(\'admin.roles.edit\', $role) }}" class="btn btn-default">Update</a>
-                            <button type="button" class="btn btn-danger delete-btn" data-destroy="{{ route(\'admin.roles.destroy\', $role) }}">Delete</button>
+                            @permission(\'roles.update\')
+                                <a href="{{ route(\'admin.roles.edit\', $role) }}" class="btn btn-default">Update</a>
+                            @endpermission
+                            @permission(\'roles.delete\')
+                                <button type="button" class="btn btn-danger delete-btn" data-destroy="{{ route(\'admin.roles.destroy\', $role) }}">Delete</button>
+                            @endpermission
                         </div>
                     ', ['role' => $role->id]);
                 })
@@ -57,13 +63,18 @@ class RoleController extends Controller
 
     public function create(): View
     {
+        validate_permission('roles.create');
+
         $role = new Role();
         $permissions = Permission::all();
+
         return view('admin.roles.create', compact('role', 'permissions'));
     }
 
     public function store(StoreRoleRequest $request): RedirectResponse
     {
+        validate_permission('roles.create');
+
         DB::transaction(function () use ($request) {
             $role = Role::create($request->only('name'));
 
@@ -79,6 +90,8 @@ class RoleController extends Controller
 
     public function edit(Role $role): View
     {
+        validate_permission('roles.update');
+
         $permissions = Permission::all();
         $rolePermissions = $role->permissions->pluck('id')->toArray();
         return view('admin.roles.edit', compact('role', 'permissions', 'rolePermissions'));
@@ -86,6 +99,8 @@ class RoleController extends Controller
 
     public function update(UpdateRoleRequest $request, Role $role): RedirectResponse
     {
+        validate_permission('roles.update');
+
         DB::transaction(function () use ($request, $role) {
             $role->update($request->only('name'));
 
@@ -101,6 +116,8 @@ class RoleController extends Controller
 
     public function destroy(Role $role): RedirectResponse
     {
+        validate_permission('roles.delete');
+
         $role->delete();
         return redirect()
             ->route('admin.roles.index')
